refactor: 웹뷰를 단일 책임 컴포넌트로 분리
- PurchaseView → PriceInputView + PurchasedLottosView로 분리
- WinningNumbersView → WinningNumberInputView + StatisticsModalView로 분리
- 각 뷰가 하나의 UI + 하나의 역할만 담당하도록 변경
